Argoverse2/requirements.txt
torch-geometric
kornia
colorama
universal-pathlib
plotly